Remove support for companion passes during fare checks (#78)
When a companion is added to a ticket, this causes fare_checker.py to silently crash without reporting an error. When a companion is added, there is no changeFlightPage link as the only way to make a change to the flight is to first cancel the companion's ticket, then make changes, and finally readd your companion.

Therefore, there will not be support for companion passes. A few reasons include needing manual intervention to acquire the flight price or canceling the companion first and checking for fares (both of which are not desired for this script).

Co-authored-by: jdholtz <joeyholtzman0507@gmail.com>
Co-authored-by: Joey Holtzman <89617856+jdholtz@users.noreply.github.com>