Don't send a notification when a reservation no longer has valid flights
A notification is no longer sent if the scheduling mechanism has already
run at least once (i.e. flights are scheduled) and a reservation that is
being monitored has all flights in the past. When this occurs, flights
will be removed from the scheduled list instead. Therefore, it is
expected that this error will occur and we don't want to notify the
user.