Fix process handling when multiprocessing start method is 'spawn'
When using the 'spawn' method, each class is pickled. However, the
Process object cannot be pickled and therefore cannot be referenced as
an instance variable.

Due to these limitations, the Process ID is now kept track of and killed
using os.kill. os.waitpid is used as well to prevent zombie processes.