Update dependencies
This removes a workaround to add the correct headless option, as that is
implemented in new versions of undetected_chromedriver. I added a couple
more workarounds to fix compatibility issues between Selenium Wire and
Undetected Chromedriver. I also added comments on the three
workarounds/extra code that can be removed once (or if...) my PRs get merged