Add FAQ question about running Docker with `--privileged`
Partially fixes #168.

This is not the ideal fix, but it has fixed people's issues of the
script getting stuck at 'Starting webdriver for current session' when
using Docker.

This issue also appears to go away for some people eventually, so the
root cause has not been identified. However, it seems platform-specific
and random at the moment (it may have to do with the network
configuration?).