Don't start checking in 5 seconds before the check-in time
I've tested this on two flights and it seems to speed up the check-in
process (at the very least, not make it slower). What usually happens is
that the request that starts 5 seconds before check-in (or, more likely,
the second or third request as the first one always fails) takes a long
time to get a response that the check-in is too early. Now, the check-in
should go through the first attempt (or maybe the second one). We will
still need to experiment with this, as delaying it 1 or 2 seconds may
make it go through sooner (succeeding on the first attempt).

This change is also done because the time is now fetched from NTP
servers, so it is a lot less likely to be out of sync with Southwest's
servers.