Always print console messages with 12-hour formats
I can add a config value later for this 'console_24_hour_time' if this
is desired. However, I'm not sure how necessary it is as the more config
values that are added, the more complex the configuration becomes. For
now, though, the default will be 12-hour time as that is what the
default for notifications is and most users likely prefer that.